Start downloading with the same function
