+++ Asspp/Interface/Download/AddDownloadView.swift
@@ -87,24 +87,14 @@ struct AddDownloadView: View {
     }
 
     func startDownload() {
-        guard var account else { return }
+        guard let account else { return }
         searchKeyFocused = false
         obtainDownloadURL = true
         Task {
             do {
-                defer { AppStore.this.save(email: account.account.email, account: account.account) }
                 let software = try await ApplePackage.Lookup.lookup(bundleID: bundleID, countryCode: account.account.store)
-                let downloadOutput = try await ApplePackage.Download.download(
-                    account: &account.account,
-                    app: software
-                )
                 let appPackage = AppStore.AppPackage(software: software)
-                let request = Downloads.this.add(request: .init(
-                    account: account,
-                    package: appPackage,
-                    downloadOutput: downloadOutput
-                ))
-                Downloads.this.resume(request: request)
+                try await dvm.startDownload(for: appPackage, accountID: account.id)
                 await MainActor.run {
                     obtainDownloadURL = false
                     hint = "Download Requested"
